fix(node): prevent integer overflow in 4 modules with wider intermediates and clamping
Addresses potential integer overflow and unsafe cast scenarios by using
wider intermediate types and bounded conversions.

- connector/incident_bundle_retention.rs: Cast total_bytes to u128 before
  percentage multiplication to prevent overflow for large storage values
- control_plane/cancellation_injection.rs: Use i128 intermediate for
  u64-to-i64 resource delta conversions to avoid wrapping on values
  exceeding i64::MAX
- connector/supervision.rs: Replace bare usize-to-u32 casts with
  u32::try_from().unwrap_or(u32::MAX) fallback to prevent truncation
  on 64-bit platforms
- supply_chain/reputation.rs: Clamp days_elapsed to i32::MAX before
  powi() to prevent panic from exponent overflow

+++ crates/franken-node/src/connector/incident_bundle_retention.rs
@@ -500,7 +500,7 @@ impl IncidentBundleStore {
         if self.max_bytes == 0 {
             return 100;
         }
-        ((self.total_bytes * 100) / self.max_bytes) as u8
+        ((self.total_bytes as u128 * 100) / self.max_bytes as u128) as u8
     }
 
     /// Check if storage is at warning level.
